04-analytics-engineering/dbt-core-big-query/ingestion/pyproject.toml
[project]
name = "ingestion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.9.0",
    "requests>=2.32.5",
]
